#39: Пофиксить предупреждения под всеми платформами и включить -Werror

+++ core/image_capture_impl.h
@@ -25,7 +25,7 @@ class ImageCaptureImpl : public ImageCapture {
       Type type,
       const std::string& source_name);
   ImageCaptureImpl(std::unique_ptr<PositionController>&& position_controller,
-      int camera_id);
+      int64_t camera_id);
   virtual ~ImageCaptureImpl();
 
   // Can return empty matrix.
@@ -33,20 +33,20 @@ class ImageCaptureImpl : public ImageCapture {
   virtual bool GrabNextImage() override;
 
   virtual double GetFocusLength() const override;
-  virtual int GetDpm() const override;
+  virtual int64_t GetDpm() const override;
 
   cv::VideoCapture* capture() const { return capture_.get(); }
   void set_focus_length(double focus_length) {
     focus_length_ = focus_length;
   }
-  void set_dpm(int dpm) {
+  void set_dpm(int64_t dpm) {
     dpm_ = dpm;
   }
 
  private:
   std::unique_ptr<cv::VideoCapture> capture_;
   double focus_length_;
-  int dpm_;
+  int64_t dpm_;
 };
 
 }  // namespace core
